refac: rename patching_strategty_types.py to patching_strategy_protocol.py
