+++ src/careamics/dataset_ng/patching_strategies/patching_strategy_protocol.py
